# packages/contracts/openapi/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media Analysis Platform API",
    "version": "2.0.0",
    "description": "Single-user channel-aware target contract for API-owned media assets, collections, immutable selection snapshots, analysis runs, artifacts, diagnostics, channel accounts, channel surfaces, and idempotency."
  },
  "paths": {
    "/v1/media-assets": {
      "post": {
        "operationId": "createMediaAsset",
        "summary": "Create one channel-account-scoped media asset without starting analysis.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/http/media-asset.schema.json#/$defs/createMediaAssetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Accepted input persisted as a media asset and visible through the inbox collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/media-asset.schema.json#/$defs/mediaAssetResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorEnvelope"
          },
          "409": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      },
      "get": {
        "operationId": "listMediaAssets",
        "summary": "List media assets for a channel account with filters, diagnostics summary, and pagination.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountId"
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "../schemas/common/enums.schema.json#/$defs/mediaKind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "../schemas/common/enums.schema.json#/$defs/mediaAssetStatus"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media assets ordered by creation time descending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/media-asset.schema.json#/$defs/mediaAssetListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media-assets/upload": {
      "post": {
        "operationId": "uploadMediaAsset",
        "summary": "Upload bytes and create one media asset without starting analysis.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "../schemas/http/media-asset.schema.json#/$defs/createMediaAssetMultipartRequest"
              },
              "encoding": {
                "metadata": {
                  "contentType": "application/json"
                },
                "file": {
                  "contentType": "application/octet-stream"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded input persisted as a media asset and visible through the inbox collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/media-asset.schema.json#/$defs/mediaAssetResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorEnvelope"
          },
          "409": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      }
    },
    "/v1/media-assets/{media_asset_id}": {
      "get": {
        "operationId": "getMediaAsset",
        "summary": "Read one media asset and safe origin or stored object metadata for the caller.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountId"
          },
          {
            "$ref": "#/components/parameters/MediaAssetId"
          }
        ],
        "responses": {
          "200": {
            "description": "Media asset detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/media-asset.schema.json#/$defs/mediaAssetResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      },
      "delete": {
        "operationId": "deleteMediaAsset",
        "summary": "Mark one media asset deleted and hide it from default collection membership reads.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountId"
          },
          {
            "$ref": "#/components/parameters/MediaAssetId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated media asset with retention metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/media-asset.schema.json#/$defs/mediaAssetResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      }
    },
    "/v1/selection-snapshots": {
      "post": {
        "operationId": "createSelectionSnapshot",
        "summary": "Create and seal an immutable selection snapshot from explicit media asset references.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/http/selection-snapshot.schema.json#/$defs/createSelectionSnapshotRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sealed selection snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/selection-snapshot.schema.json#/$defs/selectionSnapshotResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      }
    },
    "/v1/selection-snapshots/{selection_snapshot_id}": {
      "get": {
        "operationId": "getSelectionSnapshot",
        "summary": "Read an immutable selection snapshot.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountId"
          },
          {
            "$ref": "#/components/parameters/SelectionSnapshotId"
          }
        ],
        "responses": {
          "200": {
            "description": "Selection snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/http/selection-snapshot.schema.json#/$defs/selectionSnapshotResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      }
    },
    "/internal/v1/channel-accounts": {
      "get": {
        "operationId": "listChannelAccounts",
        "summary": "Internal channel identity lookup for recovery and diagnostics.",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Channel accounts ordered by last seen time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/channel-account.schema.json#/$defs/channelAccountResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "resolveChannelAccount",
        "summary": "Resolve or create a durable external channel identity for the single user.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/common/channel-account.schema.json#/$defs/resolveChannelAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved channel account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/channel-account.schema.json#/$defs/channelAccountResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created channel account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/channel-account.schema.json#/$defs/channelAccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/channel-accounts/{channel_account_id}": {
      "patch": {
        "operationId": "updateChannelAccount",
        "summary": "Update channel account status, metadata, or last-seen timestamp.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountPathId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated channel account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/channel-account.schema.json#/$defs/channelAccountResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        }
      }
    },
...
    "/internal/v1/channel-surfaces/active": {
      "get": {
        "operationId": "listActiveChannelSurfaces",
        "summary": "List active channel surfaces for restart recovery.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelAccountId"
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Active channel surfaces for restart recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/channel-account.schema.json#/$defs/channelSurfaceListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/v1/channel-surfaces/{channel_surface_id}/events": {
      "get": {
        "operationId": "listChannelSurfaceEvents",
        "summary": "List append-only channel surface event history.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelSurfaceId"
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Channel surface event history."
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/idempotencyKey"
        }
      },
...
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/pagination.schema.json#/$defs/pageCursor"
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../schemas/common/pagination.schema.json#/$defs/pageSize"
        }
      },
      "ChannelAccountId": {
        "name": "channel_account_id",
        "in": "query",
        "required": true,
        "schema": {
          "$ref": "../schemas/common/enums.schema.json#/$defs/uuid"
        }
      },
...
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Validation, channel-scope, conflict, retention, or diagnostic error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/error-envelope.schema.json"
            }
          }
        }
      }
    },
...
  }
}